// src/modules/garmentsProduction/reports/finishing/compensation-report/compensation-report-type.ts
export type CompensationReportType = {
  EFFECTIVE_DATE: string;
  COMPENSATION_NO?: string;
  ORDERPLACEMENTMONTH?: string;
  BUYER?: string;
  PONO?: string;
  ITEMTYPE?: string;
  QTY: number;
  UOM?: string;
  RATE: number;
  LOCAL_EARNING_AMT: number;
  SUPPLIER?: string;
  COMPANY_NAME?: string;
  COMPANY_ADDRESS?: string;
};